Add columnas package. Refactor freqtablesimple
All the columns classes has been implemented. Move the class freqtablesimple to freqtable.py. Some classes of the uml has been created but no implemented.
