+++ tests/test_columns.py
@@ -1,0 +1,6 @@
+import unittest
+from context import *
+
+#TODO: #2 Create tests to check columns
+class TestTablasDeFrecuencias(unittest.TestCase):
+    pass
